+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Fixed
+- Clarify instructions wrt Google Calendar (see [issue 8](https://github.com/KNowledgeOnWebScale/solid-calendar-store/issues/8))
+
 ## [0.0.14] - 2021-11-02
 
 ### Added